% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addpoly.R
\name{addpoly}
\alias{addpoly}
\title{Add polygon shading for error bars to a figure}
\usage{
addpoly(x, y, y0 = NULL, col = "grey", border = "NA", alpha = 0.5)
}
\arguments{
\item{x}{numeric of x values}

\item{y}{numeric of upper y values}

\item{y0}{numeric of lower y values. Default is zero.}

\item{col}{colour of border}

\item{border}{character with colour name or hex code}

\item{alpha}{transparency {0,1} of polygon shading}
}
\value{
Nothing
}
\description{
Add polygon shading for error bars to a figure
}
\author{
Christopher J. Brown
}
