% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_coords.R
\name{convert_coords}
\alias{convert_coords}
\title{Create a function that converts between two coordinate spaces}
\usage{
convert_coords(xyfrom, xyto)
}
\arguments{
\item{xyfrom}{vector of the form \code{c(xmin, xmax, ymin, ymax)}
giving coordinates to translate from.}

\item{xyto}{vector of the form \code{c(xmin, xmax, ymin, ymax)}
giving coordinates to translate to}
}
\value{
A function that coverts to a new coordinate space
}
\description{
This function is a closure that is used to create a function
can translate between two coordinate spaces.  Such a function
is useful for digitising graphs and for creating inset
plots.
}
\examples{
\dontrun{
# To create a function that plots a new plot ontop another
# using NDC:
posplot <- covert_coords(par('usr'), par('plt'))
}
}
\author{
Christopher J. Brown
}
