% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_resids.R
\name{ds_resids}
\alias{ds_resids}
\title{Plot Dunn-Smyth residuals}
\usage{
ds_resids(
  yobs,
  ypred,
  plotds = TRUE,
  family = "poisson",
  phis = NULL,
  trial.size = 1
)
}
\arguments{
\item{yobs}{numeric of observations}

\item{ypred}{numeric of predictions}

\item{plotds}{logical, should the residuals be plotted?}

\item{family}{character for family (poisson, binomial, negbinomial or
nbinom)}

\item{phis}{scale parameter for nbinom (= 1/shape)}
}
\value{
A Dunn-Smyth residuals and optionally, a plot of them.
}
\description{
Plot Dunn-Smyth residuals
}
\examples{
data(InsectSprays)
m1 <- glm(count ~ spray, family = "poisson", data = InsectSprays)
InsectSprays$fit <- predict(m1, type = "response")
with(InsectSprays, 
  ds_resids(count, fit, family = "poisson"))
#Looks good 

library(MASS)
m2 <- glm.nb(count ~ spray, data = InsectSprays)
InsectSprays$fit2 <- predict(m2, type = "response")
with(InsectSprays, 
  ds_resids(count, fit2, family = "nbinom", phis = 1/m2$theta))
 #Ok good, note we need 1/theta to get phi 
 
}
\author{
Christopher J. Brown
}
