% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsg_lonlat.R
\name{epsg_lonlat}
\alias{epsg_lonlat}
\title{Just returns the EPSG code for WGS84 Long-Lat cooridnates}
\usage{
epsg_lonlat(as_numeric = FALSE)
}
\arguments{
\item{none}{}
}
\value{
A \code{character} object giving "+init=epsg:4326"
}
\description{
Just returns the EPSG code for WGS84 Long-Lat cooridnates
}
\author{
Christopher J. Brown
}
