% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexalpha.R
\name{hexalpha}
\alias{hexalpha}
\title{Makes a colour value transparent}
\usage{
hexalpha(col, alpha = 1)
}
\arguments{
\item{col}{is colour name, hex value or code}

\item{alpha}{is transparency {0,1}}
}
\value{
A hex colour code
}
\description{
Makes a colour value transparent
}
\author{
Christopher J. Brown
}
