% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loccopy.R
\name{loccopy}
\alias{loccopy}
\title{Copy coordinates from a plot to clipboard}
\usage{
loccopy(n, digits = 2)
}
\arguments{
\item{n}{number of points, passed to \code{locator()}}

\item{digits}{digits to round to}
}
\value{
Sends coordinates from clicking on a plot to your clipboard
}
\description{
Copy coordinates from a plot to clipboard
}
\author{
Christopher J. Brown
}
