% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_.R
\name{paste_}
\alias{paste_}
\title{Paste with a dash}
\usage{
paste_(...)
}
\arguments{
\item{...}{one or more R objects to convert to characters}
}
\value{
A string
}
\description{
Paste with a dash
}
\author{
Christopher J. Brown
}
