% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootogram.R
\name{rootogram}
\alias{rootogram}
\title{Create a rootogram from observed and predicted counts.}
\usage{
rootogram(yobs, ypred, nbrks = 10, ...)
}
\arguments{
\item{yobs}{numeric of observed counts}

\item{ypred}{numeric of predicted counts}

\item{nbrks}{number of x-axis breaks}

\item{...}{other arguments passed to plot.}
}
\value{
A rootogram.
}
\description{
Create a rootogram from observed and predicted counts.
}
\examples{
data(InsectSprays)
m1 <- glm(count ~ spray, family = "poisson", data = InsectSprays)
InsectSprays$fit <- predict(m1, type = "response")
with(InsectSprays, 
  rootogram(count, fit, nbrks = 12))
 #A very bad fit!
 
library(MASS)
m2 <- glm.nb(count ~ spray, data = InsectSprays)
InsectSprays$fit2 <- predict(m2, type = "response")
with(InsectSprays, 
  rootogram(count, fit2, nbrks = 12))
 #Still bad!

}
\author{
Christopher J. Brown
}
