% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale2.R
\name{scale2}
\alias{scale2}
\title{Scale a variable}
\usage{
scale2(x, ...)
}
\arguments{
\item{x}{a numeric variable}

\item{...}{arguments to pass to mean() and sd()}
}
\value{
Numeric
}
\description{
Scale a variable
}
\author{
Christopher J. Brown
}
