% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semivariance.R
\name{semivariance}
\alias{semivariance}
\title{Calculate semivariance given a distance matrix}
\usage{
semivariance(xdists, yresp, ncats = NA)
}
\arguments{
\item{xdists}{a distance matrix}

\item{yresp}{response variable}

\item{ncats}{number of categories for x-axis of semivariance plot}
}
\value{
A data.frame of distances and autocorrelation statistics
}
\description{
Calculate semivariance given a distance matrix
}
\examples{
xy <- cbind(runif(50), runif(50))
y <- rnorm(50)
dmat <- as.matrix(dist(xy))
sv <- semivariance(dmat, y)
with(sv, plot(distances, semivar))

}
\author{
Christopher J. Brown
}
