% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCap.R
\name{simpleCap}
\alias{simpleCap}
\title{Change first letter of each word to upper case}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{a character}
}
\value{
x with first letter of each word in upper case
}
\description{
Change first letter of each word to upper case
}
\author{
Christopher J. Brown, from \code{?toupper}
}
