% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textspace.R
\name{textspace}
\alias{textspace}
\title{Add space between text for a label}
\usage{
textspace(x, y, labnam, nspaces = 1, ...)
}
\arguments{
\item{x}{numeric for x coordinate}

\item{y}{numeric for y coordinate}

\item{labnam}{character giving label}

\item{nspaces}{numeric giving number of spaces between letters}

\item{...}{other arguments to \code{text}}
}
\value{
Nothing
}
\description{
Add space between text for a label
}
\details{
Add spaces between text, useful for maps
}
\author{
Christopher J. Brown
}
