% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_plain.R
\name{theme_plain}
\alias{theme_plain}
\title{A plain theme for ggplot2}
\usage{
theme_plain(base_size = 12)
}
\arguments{
\item{base_size}{a \code{numeric} giving basic size of text}
}
\description{
A plain theme for ggplot2
}
\author{
Christopher J. Brown
}
