% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_groups.R
\name{with_groups}
\alias{with_groups}
\title{Evaluate expressoin by groups of grouped dataframe}
\usage{
with_groups(df, expr)
}
\arguments{
\item{df}{agrouped data frame}

\item{expr}{and expression, possible wrapped in curly braces.}
}
\value{
A list with the contents of the expression evaluated  by 
groups.
}
\description{
Evaluate expressoin by groups of grouped dataframe
}
\examples{
library(dplyr)
dat2 <- tibble(
   grp = sample(c(10, 20, 30), 100, replace = TRUE), 
   x = 1:100, 
   y = x*2 + rnorm(100) + grp)
gout <- dat2 \%>\% 
  group_by(grp) \%>\%
    with_groups(., {
     lm(y ~x)
  })
  
}
\author{
Christopher J. Brown
}
